issue found: Intermediate operation are destroyed before tensor resolution, function call overwrites, and so garbage pointer.  fix idea: CapsulatorOperation or make everything return Tensor. Second option prefered, but both can be useful, so implement both.
The issue with cut seem to originate in the fact that intermediate operations are passed as reference.
When operations are chained but not resolved in one line, those intermediate references remain as pointing to un-captured data.  This is evident by Reshape's destructor being called.

This is not an issue is: 1. operation is done in one line, whence destructor is only called after the line terminates (ie after tensor is resolved) 2. when tensor is resolved immediately after operation are held temporarily. However, the intermediateness needs to be exact, as per the second example which fails when resolution and intermediate destruction are separated by new variable declaration.

Function call is case of the second failure, where as function call is resolved, the destructed data of Reshape is overwritten and leaves garbage address in place of rTensor reference.

Therefore we have two possible resolutions:
1. make CapsulatorOperation which  will simply store value-variable of its held operation.
2. make operation holders inaccesible for inbetween usages too. This will mean we can only pass around data by tensors which could be expensive, but at the same time could be equal to CapsulatorOperation.

I suggest that second option is better for practice, but first could be a useful option to have. We will implement both.

Idea for capsulator is to call Capsule() at end of each operation when we know data will be tossed around a lot before resolution.